requirements.txt
bitstring
gettext
flask
pyyaml
lxml
protobuf3
six